Test helper to mock not finding an establishment
We want to test what happens when an establishment cannot be found on
the Academies API by it's URN, this is more likely for an academy due to
the delay in date being entered into GIAS and appearing in the academies
database.

This mock takes a urn and ensures that the response is the same shape as
a real not found result.