Show the acaedemy details on project information
Once the academy URN is provided, we attempt to fetch it's details.

If they cannot be found, likely due to the delay in GIAS > Academies
database, we show the URN and a helpful message.

If they can be found, we show them.

We have to treat fetching the establishment slightly differently to the
school, if there is no school, something has gone very wrong so we raise
an error.

The academy establish not being found is something we want to handle for
the reasons mentioned.