Create local authority model
We need to be able to show the address of the local authority to users
who are preparing the funding agreement letters. This data currently
resides in KIM and so is not sustainable.

We have been unable to locate a programatic source for this data,
so right now, our best option is to store it locally and manage it
within the service - this mirrors the exsiting process of storing the
data in KIM.