Change what MembersApi::Client.member_id returns
Originally thisd was returning a Result object in the case of error, or a
string in case of success. All other methods in this client return a Result
object. This inconsistency caused issues when trying to use this method in
the app.

Change the metho to always return a Result object, like all other methods in
the client.

Also, if the previous call to get the Constituency data fails, raise the error,
so we can handle it correctly in the application.