Show the Member of Parliament details in the UI
Add a page to the project information area, showing the information for a
school's Member of Parliament (found via searching the school's constituency
data).

If a MP is not found, show a 404 not found page.
If the Members Api is not responding, show an 500 error page.
If the constituency search returns multiple results (unlikely), show a custom
error page asking the user to contact us for help.