Fix potential name clashes in the task list
Our task attributes are used without their prefix, for example:

`redact_and_send_save`

becomes

`save`

We should try to avoid common Rails or Ruby language for this attribute
names as much as possible.

This is especially important as we begin refactoring to allow tasks to
save themselves!

We have not updated the involuntary task list as it is not used, feature
frozen and will be removed soon.