Introduce a new task form class
This is the start of a the third iteration of our task list. We want to
make it simpler to have the tasks be able to do more than update boolean
values on the task list data.

To do this the main change will be to move from the task_list updating
its own tasks, to the tasks themselves being 'first class' models that
happen to persist to one big table.

This approach is really what the form object pattern is great at and are
current task models are really following that pattern.

This first commit introduces a new `BaseTaskForm` from which all tasks
will inherit from. Much of what happens here is a repeat of the existing
tasks, but you can see that the `save` method is now on the task itself.

We've implemented one of the tasks here to test the base class - we
would have to create a sub-class to test this anyway so that feels
right.

For tasks that only have checkbox actions, all a TaskForm has to do is
define the attributes it will update - but when we need to, we can
override the save method and do anything we want.

Next we'll expose this new TaskForm on a url and add the form submission
to prove this works once intergrated.

You'll notice that `identifier` is implemented on the class and any
instance - this is to keep the API the same as the existing tasks and
should be resolved over the course of the refactor.