Introduce new tasks controller
We have created a new TaskForm object that can save itself, our next
step is to allow one of these TaskForms to be rendered.

This work introduces a new TasksController that is capable of rendering
any TaskForm, to validate this we have rendered the redact and send task
at:

`new_tasks/<project_id>/tasks/redact_and_send`

Right now, we just want to validate everything is working as we would
expect so some of this work might seem a bit off, mostly around what we
do when a TaskForm is saved, the expectation is to go back to the task
list, which we will eventually, but not here.

Because this controller can handle any TaskForm, it has to work quite
hard. It has to be able to convert a path into the correct class and
render the correct view. We could move this into some kind of factory
object, but for this stage of the refactor, I think this makes the
review a little more straight forward.

None of this is intended to be seen by users.