Fix subtle bug with URN validation on project form
A very subtle bug occurred with the CreateProject form, allowing two projects
with the same URN & UKPRN combination to be created.

This heppened because the duplicate project had a nil `assigned_to` value, which
meant it did not appear in the `Project.in_progress` scope.

Adding a new `Project.not_completed` scope to use instead for this validation
has fixed the issue. A project which has no `assigned_to` value is not considered
`in_progress` because no-one will actively be working on it.