Introduce the funding agreement contact task
This is part of our refactoring of the task list, here we want to prove
the new TaskForms can handle the case where there are no attributes that
are saved in task_list table via the task_list.

In current version this would be very, very difficult because the
task_list saves the task and would assume the attributes are all in the
task_list table.

We used this new 'funding agreement contact' task to prove this as it
creates a new Contact model. The final design and content has yet to be
delivered for this so don't take this work as the final version of this,
we simply want to validate the models can handle this case.

This work demonstrates how we can override the initialize and save
methods on a TaskForm and do whatever we like.

Here, once the TaskForm is initialized, we use find_or_create_by to
locate any existing main contact.

We validate the form as normal and either create the appropriate contact
or update the existing one.

As a feature, this main contact is not finished, but as a proof, it
works very well.

Worth noting you cannot use shoulda matcher to test these FormTasks
because of how they are initialized.

This work is not user facing and will be refined before users see it.