List of projects added by a user
When a regional delivery officer creates a new project and hands it over
to regional casework services, it dissappears from their view. This
means it is difficult for them to find it again and add notes or
contacts.

We already know the projects a user has 'added' so we can help by
showing a list of these projects.