Create ConversionDatable form concern
Concerns are a Rails design pattern that exist for other kinds of
objects.

https://api.rubyonrails.org/classes/ActiveSupport/Concern.html

They are a little contraversional being a DHH thing, especially the
whole 'able' naming convention.

That said, we have some and they work well. I am happy to at least get a
feel for them so here is one for our form objects.

The idea is to encapsulate all the requirements for handling a
multiparameter date, here the confirmed conversion date, in a concern,
you include the concern where a form is:

> concerned with the conversion date

this does two things:

- encapsulates all the behaviour
- allows it to be reused

at the cost of:

- some misdirection
- some 'able' naming

I feel like the 'able' naming breaks down the misdirection a little
because if you see anything 'able' I hope you realise it is a concern -
they are just slightly fancy modules after all.

This concern supplies the object that includes it with all the things it
needs to deal with a converison date submitted via the dreaded GOV.UK
multipart date field.