Projects fetch their provisional conversion date
We are moving to a world where we will not need to store any version of
the conversion date other than the current date and the history of
changes.

In this world, the provisional conversion date can be modelled as:

> Get all the conversion date histories ordered by created at and retrun
the first. The `previous_date` will be the provisional conversion date.

Here we add a method to Conversion::Project to fetch the provisional
converison date based on this logic, we prefix the method with fetch to
prevent name colisions.