Add the Stakeholder kick off task
This is the thrid and most complex proof that the new task model can
handle everything we want it to be capable of.

This is the task where the user is expected to confirm the conversion
date by supplying it as part of the task. This works now via a bunch of
not very pretty hacks.

Here we demonstrate how this new model can handle this use case, as ever
we overide the initializer and save methods with the custom behaviour.

Unfortunately, for now we have to do extra work to keep the old task
list happy - this can be stripped out later.

As you can see we take on responsibility for creating a
ConversionDateUpdater and calling it, this creates the correct date
history and stores the previous state, this is not possible with the
current models.

As mentioned there is no need to store or update an attribute on the
TaskList table, but we have to do it whilst we need to support the old
models along with the new.