Add the new task list model
Now that we have created 'tasks' that can exists without a list, it is
time to wrap them up in a simple list so they can be displayed together.

The goal is to re-create the existing task list as closely as possible.

It follows a similar path as the tasks themselves, there is a base class
that does eveything except define the tasks which make up the list,
you'll see the works exactly as it did before.

The API of the task list model is intentionally the same as the existing
one so very little needs to be done to make things work, mostly
duplicating views in the new locations.

We have made a couple of methods very intentionally class methods as
they provide the same response for any instance of a task list.