Add AZURE_REDIRECT_URI env var
In 901c1d4587d56c64c4027b7f03052d80fcd55439 we added a new AZURE_REDIRECT_URI
var to support that the app is now behind a proxy on Azure.

This env var is required for the app to run locally. Add it to the Dotenv
required keys, and dev & test env example files.