Add optional tasks
Building on the refactored tasks and task list, here is the optianal
task. Optional tasks are those that can be marked as 'not applicable'
and will be considered not required to complete the project.

Here we add the BaseOptionalTaskForm which provides the behaviour, any
optional task inherits from this class to add the behaviour.

We test this by adding an optional task, Articles of assication.