Correctly enforce project ID constraints
Previously, visiting `/projects/:id` with an incorrect project ID was correctly
rendering a 404, while visiting `/conversions/voluntary/projects/:id/task-list`
was incorrectly rendering a 500 error.

This was because we were only applying the route constraint on the "outermost"
`projects` resource group. Within the voluntary & involuntary nameapaces
we were only applying the constraint to the `#show` route. So visiting any
of the concern routes with an incorrect project ID made the application pass
the ID to the database and throw an `ActiveRecord::StatementInvalid` error
(rendered as a 500).

Encapsulate all the namespaced routes with the routing constraint. This means
the `#show` page *and* all the concernable pages will be subject to the route
constraint.