Try to capture exceptions in Applcation Insights
We have been asked to use MS Application Insights if possible.

We are already sending Rack telemetry, here we are seeing what exception
handling might look like.

We are using the archived gem to achieve this right now, it may support
all that we need we do not really know.

According to the docs, this is all that is required:

https://github.com/microsoft/ApplicationInsights-Ruby#collecting-unhandled-exceptions

I've tested this locally and it sent the exception I raised manually so
this appears to be a good first step.

The plan is to deploy this and see if any exceptions are raised in our
existing service (Sentry) and from here, also in Application Insights.