Stop ignoring bundler config and configure it
We have seen multiple instances of developers struggling to build tiny
tds because of the free-tds dependancy.

It takes us a while but we usually figure out that it is and define this
environment variable.

I cannot think of a reason to not have some local bundler config so
here we remove it from gitignore and commit the thing we keep
forgetting.

There is a small chance developers not using Brew will have to deal with
this - but those users are far less likely than those that will benefit
from this change.