Add new Conversion Task generator
As with the old task list, we can encapsulate what is requried to add
new Conversion Task with a simple generator to add the necessary files.

As the new task model is different to the old, we cannot reuse the
shared example, creating a simple spec instead.

We decided that checking the resulting files for specific strings would
be enough rather than diffing the entire file.