Display local authorities and their details
We are displaying the local authorities as there will be a requirement
from a service support role to be able to add, edit and delete any local
authorities within the service.

This is just the first stage in getting the local authority data up and
displayed. Next will introduce the service support role and set access
permissions the create, update and delete methods.