Users can edit the Academy URN on a project
Users need a way to add the new Academy URN to a project. From the
All conversions without an academy URN" page, the user can click on a link
to a form to add just the Academy URN.

When they fill in the Academy URN on the form and click "Save and return", the
URN will be checked in GIAS. If it is found, the user can choose to confirm or
cancel depending on whether or not the URN is correct.

If the URN is not found, they can choose to re-enter the URN, or cancel.

If they do not supply a URN and submit the form, they see an error message.

If the Academies API times out, they see a helpful error message.