Remove Involuntary conversion specs
We are no longer having a seperate involuntary conversion model, it is
simply an attribute of a conversion, it is also now called 'sponsored'.

Our first stept to removing the related code is to delete the specs,
this will cause the test suite to go red due to the lack of coverage,
but this should get resolved in the next commits.