Remove two projects scopes
These two scopes are now served by other scopes:

`voluntary` and `sponsored`

We may have missed some other places but this work is part of a larger
set of commits that remove all of the involuntary converison code and so
will get resolved in later commits.

We have to update the Brakeman ignore list for these two - they are the
same calls, just with new scopes.