Continue to rationalise the routes
Here we remove all the old dynamic routes and introduce a new 'base'
resource 'projects' - we've decide to start this work by assuming all
projects are one type, which is the case right now.

This breaks eveything, but gives us something to work back to a working
system in the next commits.