Switch the task list routing concern
This the first sight of the new task list models in use.

Here we create a new concern for the new version, it uses a 'tasks' path
because that is a better reflection of the models and resource routing -
the task list is a list of all Task objects after all.