Update TaskList model
As promised some time ago now, the TaskList model becomes very simple,
it is really only used to have a single table to back all the tasks in
the database.

This works removes most of the code, leaving just enough to keep the
test suite happy.

We will come back and refactor this later, for now our focus is to ship
the new task list so this will have to do.