Add TasksData
The new tasks model splits the responsibility of storing the tasks data
and representing the task list into two:

- TasksData is an ActiveRecord object that represents the database
  table and does almost nothing
- TaskList represnets the task list, which tasks and the order

Here we introduce the new TasksData model and delegated type that will
replace the existing TaskList. We switch out the existing TaskList for
the new TasksData - this will break lots of things, which will get fixed
next.