Fix url helper for mailer
This probably doesn't need to exist, it certainly doesn't beed to check
the task list type, since the new routes were introduced, this will
always be `project_url`.

We'll leave it in for now as we are focused on keeping this work as
small as we can to introduce the new task list.