Fix: Stakeholder kick off task validation
We validate at the form submission so blank fields on the form come to
us as empty strings. We convert the string to an integer, and in Ruby ""
= 0 so the test for `month` and `year` `blank?` was never true, it was
0.

Here we alter th  guard inside the methods to check for empty strings or
nil - this is exactly what `blank?` does!

This and gives the desired behaviour: the form can be submitted without
a confirmed conversion date.