Introduce Single Table Inheritance for Contacts
We wish to add a new "type" of contact - Director of Child Services. In order to
do this, we must first add Single Table Inheritance to the Contact model, and
change the existing Contact model into a Contact::Project model (contacts
associated with projects).

Adding STI to the Contacts involves adding a `type` column to the model, and
moving the existing Contact code into Contact::Project. We will then be able
to add other Contact types later.