Remove shared example for conversion project controller
Now that we only have a conversio project, we don't need to share the
behaviour in the specs.

Here we bring the request specs into the same spec file. We also add
coverage around the behaviour depending on if the project is assigend to
regional casework services or not, covering this in a request spec means
we can remove the converage in the slower feature spec.