Delete the ProjectConversionDateMigrator
We used this to migrate the provisional conversion date of projects, the
rationale is we wanted the result of this 'data migration' to be tested.

The migration that used this class has long since been run in production
and should not be run again.

In other environments, the schema is loaded rather than stepping through
each migration file.

This class is becoming difficult to maintain as we move away from the
world where it help us.