Display trust's group identifier within the project information view
Although the Acadmies API only lets us search for trusts by UKPRN, we
have learnt that users prefer other identifiers in their searches.
We have identified that Group Identifier (ID) is one of their
preferences and one that we are able to access via the Academies API. So
to help users gain more confidence that they have the correct trust, we
are displaying this group identifier within the project information
view.