Remove stakeholder_kick_off_confirmed_conversion_date
This attribute was needed to support the old task list, which is now
gone so we can remove it.

The task this attribute belongs to included extra code to keep the old
task list happy, this can also be removed now and a slight change to
make it work for only the new task list.

The guards sprinkled across this are not ideal, the status of the task
now relies on an attribute of `project`, I am not sure how to avoid
this.