Conversion date history notes cannot be deleted
Changing the conversion date is a signigicant event and we do not want
to loose the note associsated with the change.

We use the Note model to store it, which by design, allows the user who
created the note to delete it.

This change prevents that for notes associated with a conversion date
history change.