Show the academy name to service support users
Once the academy name is confirmed, a service support users adds the new
record to Get infromation about schools, we then read that information
back in and show it.

So the service support users knows the academy name and can provide it
we show it on our views that support this process.

Once the new academy record is created the service support user adds the
URN and can verify the correct name is being returns - for this reason,
the 'with academy URN' view shows the fetched name NOT the local
'confirmed' name.