Switch to MOJ style header
We are an internal service and not part of GOV.UK, we should not use all
aspects of the GOV.UK design system, though we should use most of it.

Here we load all of the MOJ design system and switch the GOV.UK header
for a more department specific one.

We expect to have to switch to the DfE design system at some point, but
this seems like a simple first step.

We load the MOJ styles like this becuase we cannot use the standard
`@import all` because the MOJ system iteself tries to include the
govuk-frontend, which will not work in our setup.