Remove signed in users email verification text from the view
We now do not need to display to the user who is signed in from within
the service. We had added this in the early stages of the product as a
way to validate the sign in was working.

This commit also updates the existing sign in specs to reflect this
removal