Display generated DfE number
The DfE number is another identifier for the establishment which may be
useful for users to use as a form of validation.
It is also known as LAESTAB.

The DfE number is shown within the GIAS but isn't accessible via the
AcademiesApi, so we have therefore needed to construct it from the local
authority code and the establishment number.