Add the Funding Agreement Letters task
This task was started by @mec but I have changed the implementation.

Instead of a new Project::Contact being added to the project as part of the
task, instead the user is presented with a list of existing contacts on the
task page and can choose one as the Funding Agreement Letters contact. This
contact's ID is then associated with the project as its funding agreement
contact.

If there are no contacts on a project, the user is directed to add some at the
existing External Contacts page.

At this point the user is able to change the Funding Agreement Letters contact
if one has already been chosen, but I'm not entirely sure this is required.