Fetch and expose establishment diocese code
Right now the only marker we have if a school is considered 'religious'
is if it has a diocese. When no diocese is present the Academise API
returns the string 'Not Applicable'.

We decided not to use the string and instead fetch the 'diocese code'
which seems to be consistenty '0000' when the diocese is not applicable.

There are other markers on the academuies API that might be better, but
we could see no consistency to the codes, so decided to use this until
there is something better.

The method is not used in this commit, that comes next.