New project logic for church funding agreement
If a school converting to an acadmey is not considered 'religious' the
Church supplemental agreement is likely to be considered not applicable.

Here we add that logic to the new conversion project form, setting the
appropriate value as needed.

Users can still set the task back to applicable as desired.