New project logic sponsored support grant task
When a new project is created that has not been issued a directive
academy order i.e. the project is considered 'voluntary' the 'Process
the sponsored support grant' is likely to not be applicable as only
'sponsored' projects, those issued with a directive academy order,
receive the sponsored support grant.

Here we add this logic into the create project from object.