Add method to MembersApi client
The members API client has to send a number of request to get the data
we are interested in. Instead of repeating those steps anywhere we need
to use MP data, instead we have added a method that collects them into
one call.

The method `member_for_constituency` takes an Establishment's
`parliamentary_constituency` and returns the MP details as we need them
in a single hash.