Add policy for header navigation
We want the header navigation to respond to the role of the user that is
signed in, showing or hiding navigation items as appropriate.

Here we leverage a policy class to define this, in use we can add:

`if policy(:navigation).show_header_navigation_item?`, etc

to each navigation item.