(Fix) Redirect the user if they visit `/projects`
Going to `/projects` was resulting in a 500 error because the app did not know
what to do with that route. The route isn't part of any navigation but it is
not inconceivable for a user to visit the path.

Add in redirection to take a lost user to the all in-progress projects page.