Add header navigation
Whilst we have not fully fleshed out all the navigation for the site, we
really need to be getting users access to all the features of the
application.

We are confident in certain groupings of projects and that we will start
things off by using the MOJ design system ahead of the DfE one, which is
not in a fit state right now.

This first level of navigation is shown in the header, the MOJ header
allows this, but does not seemt to have a visual 'active' state that we
need, so we have added a basic one, this can be refined.

The visual treatment we have used is used by the GOV.UK header
navigation.

We only want certain navigation items to show up for certain user roles,
this is handled here by a navigation policy.

See more about the MOJ header:

https://design-patterns.service.justice.gov.uk/components/header/

And the GOVUK one it is based on:

https://design-system.service.gov.uk/components/header/