Rationalise two service support views
The two views we have to help service support users manage the new
academy name for a conversion are not in the right place.

Here we bring the two routes together under a
`/projects/service-support` path and deal with the resulting changes.

We end up with the paths:

`/projects/service-support/with-academy-urn`
`/projects/service-support/without-academy-urn`

and related controllers and views.

Nothing changes in the views, only the paths, helpers, locations, etc.