Add basic primary navigation
Now we have the top level 'header' navigation, we want to introduce the
basics of the second level 'primary' - we are using the MOJ style and
naming here.

For our existing 'All Projects' header navigation, we add a primary
navigation, preparing the ground for other primary navigation to match
up to all the header navigation 'sectsions'.

With this change, all users can now move between:

- all in progress projects
- all completed projects
- all project statistics