Introduce new opening routes and update
The 'openers' and 'revised conversion date' views work as a pair.

Here we introduce a rationalised pair of paths to back that more
explicit. Bringing these two view together in this way means the
routing matches up with the views and works along with the new primary
navigation to help orient users.

We also supply default month and year values at the routing level, this
will simplify some code later once we swap to use the new routes.

We then switch to use the new routes and add the 'Opening' section to
the primary navigation.

Like the header navigation, we have to have the concept of a 'namespace'
so that the parent navigation can remain 'active'. I am confident there
is a better way to manage this, but I think this will work for right
now.