Update statistics page layout
Now that users can navigate to this page themselves, it stands out as
having type sizes that are not consitent with the rest of the views that
feel part of this section.

Here we adjust the display sizes, but no the markup, to bring it inline.

We also loosely want the page title to match the navigation where we
can, so we dropped the 'page'.