Create an opening projects csv exporter
Users need to export the projects in the openers list to allow them to
populate the funding agreement letters.

We need a way to go from a collection of projects to a csv file.

Here we kick off that work with a class that covers off some basic
attributes of a project.

We want to set the right path to follow as there are at least 44
attributes that will end up in the csv which means a lot of repetition,
so we want a solid foundation to work from.

Our goal is to validate we can populate the csv with all the right data
from the various sources. We can then add additional attributes with
confidence.