Fix: redirect users with no role
We will soon have the potential of users without a role. We want them to
see the All projects section, but have no header navigation (becuase
there would be no other option).

Before this change, users without a role when redirected the the wrong
location.

Here we fix that and update the feature specs around the 'root' page
redirection.