Create a simple MP object
Our Members (of Parliment) API wrapper models the actual responses from
the API which are varied.

In previous work we created a single method to go from a
Project.establishment.parlimentary_constitunency > Member of Parliment
details, which retruned a hash.

This work takes that further by representing the MP Details the way we
want and so makes the interface simpler.

We will use this object soon.