Update OpeningProjectsCsvExporter
Previously we added a new MemberDetails to the MembersApi, this returns
a simple object as we want it so we can interact simply with it.

Here we update the OpeningProjectsCsvExporter to use it.

We also move the column headers to the locales and make it so that the
columns names and order only need to be defined once.

This does mean we MUST recieve at least one Project now, which is a
shame, as we have to handle that, but the benefit is that there is no
way to get the order or position of a value wrong.