Add Trust name
We have a lot of attributes to add to the opening project export, here
we add trust name so that we can test real world performance as adding
this will force the project to call the Academies API to fetch the trust
details.

This should be the last thing to add to get a sense of the performance
of creating the export file, if we have done things well, we will only
make one call to the APIs per entity, per row.