Clean up sign in spec and helper
The sign in is about to become a little more complex, here we take some
time to clean up both the spec, helper and controller.

The main difference is we add the 'groups' to the Omniauth details we
get back upon successful authentication - this is in the real response
and will be used soon.

To keep everything happy, we supply a default set of groups only
overridding when needed.