Add and collect a Users Active Directory group ids
We want the application to authorise users without the team having to do
anything. We also want the application to understand which region a user
is considered to be part of.

Here we add a column to users to store the group ids that are returned
upon successful authentication.

Everytime a session is created, we update the groups incase they have
changed.

We keep things simple by encoding the array of group ids into a single
column as a user will belong to far more groups than we model.