Different provisional/confirmed conversion dates
Previously we stored the provvisional covnersion date in a seperate
column. We then removed that and simply mark the project as 'with
provisional conversion date'.

However, we missed the impact this has on the
`self.conversion_date_revised_from` scope. Now the first Converison Date
History  is always the change from provisional to confirmed and, if the
dates are diffrent, the test is triggered and the project included when
it should not be.

The fix is to exclude any project that only has one
Conversion::DateHistory when grouping. A project in this state must only
have been confirmed and all projects must be confirmed.