Rescue error when an MPs contact details are nil
Sentry: https://sdd-n7.sentry.io/issues/4267882580/?project=6684508&referrer=slack

We received some errors indicating that the contact details object for an MP
were nil. We're not sure if this is happening because the MP details are not in
the API, or if we're being rate-limited, or some other reason.

If this happens, rescue the NoMethodError and throw a more informative error,
so we can begin to debug this in production.