Service builds a set of trusts with project counts
We want to display a list of the trust that are represented by the in
progress projects.

As this is a little complex, we wrap it in a service object.

The goal is to return a simple set of trust details along with the count
of projects, right now, only conversion which can be used to render a
view for users to choose a trust to view.

This way we only render trusts that have projects.

Testing is a bit ugly, but like most services we only test the command
method, here `call`, stubbing all the necessary trust details is pretty
messy, so we kept eveything to one spec as the output is pretty clear.

We did try to inject the projects as a dependancy, but this didn't make
the test much cleaner as it is really the trusts of projects that we
need and those are fetched from the API.

Next we can use this service to build the view.