Add by trust view
Users want to see all the projects in the application by the trust they
are interacting with.

We previously built a service to provide a scope of these projects, here
we use the service to build a view.

Users come to the by trust section and are shown a list of all the trust
for the projects that are not completed. They can click a link to view
all the projects for the trust represented by that row.

The show view itself alrady exists, so we use it, whilst it is no
strictly a `show` it is not an `index` either - maybe?