Page the by trust view
In most cases there will be at most one trust for evey project that is
in progress, we already have ~150 projects so we can expect the number
of trusts shown in this view to be near that.

For this reason we turn on paging. To do this we need to enable the
'array' extra from the Pagy gem:

https://ddnexus.github.io/pagy/docs/extras/array/

as the service returns an array not an ActiveRecord collection.