Service to fetch projects by local authority
Some users are 'Local authority relationship managers' and want to see
the projects in a given local authority.

Here we build a service to wrap the steps invloved, with the aim of
returning a list of local authorities that have projects in the
applicaiton.

First we fetch all projects that are not complete and group them by
their `establishment.local_authority_code`.

We then fetch the local authorities useing the codes, loop over them
counting the projects as we go.