Add pager to local authorities with project view
As ever, it doesn't feel like the amount of stubbing we would have to do
to test the pager is worth the confidence it would bring over knowing
Pagy (the gem we use) works as intended.

It would be nice to thing there is a way to test this, but we have yet
to finding beyond creating enought unique local authority and projects
(in this case) to trigger paging.

The motiviation for paging comes from the likelyhood that the number of
local authorities could be over the 20 row limit we set ourselves for
tables.