Add a view of regions with projects
We use the new ByRegionProjectFetcherService to get the number of
projects in a region and display them.

As there are only nine regions we don't need a pager.

We could show wll regions and 0 values, but all the other 'by something'
views work like this, so this is consistent, we can get feedback.