Add a by user project fetcher service
This is a pattern we have now, a service that fetches projects and
groups them by user (here) counting up the projects and creating little
simple objects to render in a view.

This one could return user objects, but it feels neat to include the
conversion project count and only what is needed in a OpenStruct.