Add paging to view of users with projects
It is likely for the number of users in this view to be high, and at
least above our self imposed 20 row max for a table.

The actual list of projects feels like 20 is plenty so we won't page
that view without evidence.

It's still frustrating there is no straightforward way to test this, but
that is the precedence that has been set - and I cannot find a good way.