Remove columns from opening view
This view was meant to focus on acadmey operations user needs but
actually is more generic than that. It is more valuable to sell the
service design to users that this view replaces the 'baseline'
documents.

The two columns removed only speak to acadmey operations user and are
surpless to requirements for other users.

Acadmey operations users will get their own version of this view at some
point later, but for now we remove them.