Update EstablishmentsFetcher
I am not in love with this service thing, it has a massive side effect
of 'injecting' a bunch of projects with establishment, which is great,
but it is not obvious what it is doing in use.

That said, I don't have capacity to address that and it does the job.

What we do need is for it to handle high numbers of projects, right now
it is pretty safe to use because we happen to call it on paged views
with a max of 20 projects and the API bulk endpoint seems happy with
that.

We don't know what the upper limit would be though, can we request
establishments for hundreds of projects at once?

It feels reasonable to batch these higher numbers regardless, so here
we introduce batching to the object. We use Rails `in_batches` to help
us and relies on ActiveRecord to work, so we make sure the projects
passed in are the result of some kind of ActiveRecord query
(ActiveRecord::Relation).

I found the test hard to parse, so refactored it - it is a bit more
verbose, but it let me work through the problem.

The goal is that the user of EstablishmentsFetcher should not worry
about how many projects are passed in, from 1 to 1000, it should handle
it.